chore: refactor the Clusters page into smaller components
