+++ src/app/Clusters/Clusters.tsx
@@ -1,554 +1,8 @@
-import {
-  PageSection,
-  PageSectionVariants,
-  Panel,
-  ToolbarItem,
-  Toolbar,
-  ToolbarContent,
-  TextContent,
-  Text,
-  SearchInput,
-  Spinner,
-  MenuToggle,
-  MenuContent,
-  MenuList,
-  MenuItem,
-  Menu,
-  Popper,
-  ToolbarToggleGroup,
-  ToolbarGroup,
-  ToolbarFilter,
-  Badge,
-} from '@patternfly/react-core';
-import { renderStatusLabel } from 'src/app/utils/renderStatusLabel';
-import { Table, Thead, Tr, Th, Tbody, Td, ThProps } from '@patternfly/react-table';
-import React, { useEffect, useState } from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { getClusters } from '../services/api';
-import { Cluster } from '@app/types/types';
-import { FilterIcon } from '@patternfly/react-icons';
-interface TableToolbarProps {
-  setSearchValue: (value: string) => void;
-  searchValue: string;
-  setStatusSelection: (value: string) => void;
-  statusSelection: string;
-  setProviderSelections: (value: string[]) => void;
-  providerSelections: string[];
-  onSearchChange: (value: string) => void;
-}
-
-const TableToolbar: React.FunctionComponent<TableToolbarProps> = ({
-  searchValue,
-  setSearchValue,
-  setStatusSelection,
-  statusSelection,
-  setProviderSelections,
-  providerSelections,
-}) => {
-  const onSearchChange = (value: string) => {
-    setSearchValue(value);
-  };
-
-  // Set up name search input
-  const searchInput = (
-    <SearchInput
-      placeholder="Filter by account name"
-      value={searchValue}
-      onChange={(_event, value) => onSearchChange(value)}
-      onClear={() => onSearchChange('')}
-    />
-  );
-
-  // Set up name input
-  const [isStatusMenuOpen, setIsStatusMenuOpen] = React.useState<boolean>(false);
-  const statusToggleRef = React.useRef<HTMLButtonElement>(null);
-  const statusMenuRef = React.useRef<HTMLDivElement>(null);
-  const statusContainerRef = React.useRef<HTMLDivElement>(null);
-  const handleStatusMenuKeys = (event: KeyboardEvent) => {
-    if (isStatusMenuOpen && statusMenuRef.current?.contains(event.target as Node)) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsStatusMenuOpen(!isStatusMenuOpen);
-        statusToggleRef.current?.focus();
-      }
-    }
-  };
-
-  const handleStatusClickOutside = (event: MouseEvent) => {
-    if (isStatusMenuOpen && !statusMenuRef.current?.contains(event.target as Node)) {
-      setIsStatusMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleStatusMenuKeys);
-    window.addEventListener('click', handleStatusClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleStatusMenuKeys);
-      window.removeEventListener('click', handleStatusClickOutside);
-    };
-  }, [isStatusMenuOpen, statusMenuRef]);
-
-  const onStatusToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation();
-    setTimeout(() => {
-      if (statusMenuRef.current) {
-        const firstElement = statusMenuRef.current.querySelector('li > button:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
-    setIsStatusMenuOpen(!isStatusMenuOpen);
-  };
-
-  function onStatusSelect(event: React.MouseEvent | undefined, itemId: string | number | undefined) {
-    if (typeof itemId === 'undefined') {
-      return;
-    }
-
-    setStatusSelection(itemId.toString());
-    setIsStatusMenuOpen(!isStatusMenuOpen);
-  }
-
-  const statusToggle = (
-    <MenuToggle
-      ref={statusToggleRef}
-      onClick={onStatusToggleClick}
-      isExpanded={isStatusMenuOpen}
-      style={
-        {
-          width: '200px',
-        } as React.CSSProperties
-      }
-    >
-      Filter by status
-    </MenuToggle>
-  );
-
-  const statusMenu = (
-    <Menu ref={statusMenuRef} id="attribute-search-status-menu" onSelect={onStatusSelect} selected={statusSelection}>
-      <MenuContent>
-        <MenuList>
-          <MenuItem itemId="Unknown">Unknown</MenuItem>
-          <MenuItem itemId="Running">Running</MenuItem>
-          <MenuItem itemId="Stopped">Stopped</MenuItem>
-        </MenuList>
-      </MenuContent>
-    </Menu>
-  );
-
-  const statusSelect = (
-    <div ref={statusContainerRef}>
-      <Popper
-        trigger={statusToggle}
-        triggerRef={statusToggleRef}
-        popper={statusMenu}
-        popperRef={statusMenuRef}
-        appendTo={statusContainerRef.current || undefined}
-        isVisible={isStatusMenuOpen}
-      />
-    </div>
-  );
-
-  // Set up provider input
-  const [isProviderMenuOpen, setIsProviderMenuOpen] = React.useState<boolean>(false);
-  const providerToggleRef = React.useRef<HTMLButtonElement>(null);
-  const providerMenuRef = React.useRef<HTMLDivElement>(null);
-  const providerContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const handleProviderMenuKeys = (event: KeyboardEvent) => {
-    if (isProviderMenuOpen && providerMenuRef.current?.contains(event.target as Node)) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsProviderMenuOpen(!isProviderMenuOpen);
-        providerToggleRef.current?.focus();
-      }
-    }
-  };
-
-  const handleProviderClickOutside = (event: MouseEvent) => {
-    if (isProviderMenuOpen && !providerMenuRef.current?.contains(event.target as Node)) {
-      setIsProviderMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleProviderMenuKeys);
-    window.addEventListener('click', handleProviderClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleProviderMenuKeys);
-      window.removeEventListener('click', handleProviderClickOutside);
-    };
-  }, [isProviderMenuOpen, providerMenuRef]);
-
-  const onProviderMenuToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation();
-    setTimeout(() => {
-      if (providerMenuRef.current) {
-        const firstElement = providerMenuRef.current.querySelector('li > button:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
-    setIsProviderMenuOpen(!isProviderMenuOpen);
-  };
-
-  function onProviderMenuSelect(event: React.MouseEvent | undefined, itemId: string | number | undefined) {
-    if (typeof itemId === 'undefined') {
-      return;
-    }
-
-    const itemStr = itemId.toString();
-
-    setProviderSelections(
-      providerSelections.includes(itemStr)
-        ? providerSelections.filter((selection) => selection !== itemStr)
-        : [itemStr, ...providerSelections]
-    );
-  }
-
-  const providerToggle = (
-    <MenuToggle
-      ref={providerToggleRef}
-      onClick={onProviderMenuToggleClick}
-      isExpanded={isProviderMenuOpen}
-      {...(providerSelections.length > 0 && {
-        badge: <Badge isRead>{providerSelections.length}</Badge>,
-      })}
-      style={
-        {
-          width: '200px',
-        } as React.CSSProperties
-      }
-    >
-      Filter by provider
-    </MenuToggle>
-  );
-
-  const providerMenu = (
-    <Menu
-      ref={providerMenuRef}
-      id="attribute-search-provider-menu"
-      onSelect={onProviderMenuSelect}
-      selected={providerSelections}
-    >
-      <MenuContent>
-        <MenuList>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('AWS')} itemId="AWS">
-            AWS
-          </MenuItem>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('GCP')} itemId="GCP">
-            Google Cloud
-          </MenuItem>
-          <MenuItem hasCheckbox isSelected={providerSelections.includes('Azure')} itemId="Azure">
-            Azure
-          </MenuItem>
-        </MenuList>
-      </MenuContent>
-    </Menu>
-  );
-
-  const providerSelect = (
-    <div ref={providerContainerRef}>
-      <Popper
-        trigger={providerToggle}
-        triggerRef={providerToggleRef}
-        popper={providerMenu}
-        popperRef={providerMenuRef}
-        appendTo={providerContainerRef.current || undefined}
-        isVisible={isProviderMenuOpen}
-      />
-    </div>
-  );
-
-  // Set up attribute selector
-  const [activeAttributeMenu, setActiveAttributeMenu] = React.useState<'Account' | 'Status' | 'Provider'>('Account');
-  const [isAttributeMenuOpen, setIsAttributeMenuOpen] = React.useState(false);
-  const attributeToggleRef = React.useRef<HTMLButtonElement>(null);
-  const attributeMenuRef = React.useRef<HTMLDivElement>(null);
-  const attributeContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const handleAttribueMenuKeys = (event: KeyboardEvent) => {
-    if (!isAttributeMenuOpen) {
-      return;
-    }
-    if (
-      attributeMenuRef.current?.contains(event.target as Node) ||
-      attributeToggleRef.current?.contains(event.target as Node)
-    ) {
-      if (event.key === 'Escape' || event.key === 'Tab') {
-        setIsAttributeMenuOpen(!isAttributeMenuOpen);
-        attributeToggleRef.current?.focus();
-      }
-    }
-  };
-  const handleAttributeClickOutside = (event: MouseEvent) => {
-    if (isAttributeMenuOpen && !attributeMenuRef.current?.contains(event.target as Node)) {
-      setIsAttributeMenuOpen(false);
-    }
-  };
-
-  React.useEffect(() => {
-    window.addEventListener('keydown', handleAttribueMenuKeys);
-    window.addEventListener('click', handleAttributeClickOutside);
-    return () => {
-      window.removeEventListener('keydown', handleAttribueMenuKeys);
-      window.removeEventListener('click', handleAttributeClickOutside);
-    };
-  }, [isAttributeMenuOpen, attributeMenuRef]);
-  const onAttributeToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation(); // Stop handleClickOutside from handling
-    setTimeout(() => {
-      if (attributeMenuRef.current) {
-        const firstElement = attributeMenuRef.current.querySelector('li > button:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
-    setIsAttributeMenuOpen(!isAttributeMenuOpen);
-  };
-
-  const attributeToggle = (
-    <MenuToggle
-      ref={attributeToggleRef}
-      onClick={onAttributeToggleClick}
-      isExpanded={isAttributeMenuOpen}
-      icon={<FilterIcon />}
-    >
-      {activeAttributeMenu}
-    </MenuToggle>
-  );
-  const attributeMenu = (
-    // eslint-disable-next-line no-console
-    <Menu
-      ref={attributeMenuRef}
-      onSelect={(_ev, itemId) => {
-        setActiveAttributeMenu(itemId?.toString() as 'Account' | 'Status' | 'Provider');
-        setIsAttributeMenuOpen(!isAttributeMenuOpen);
-      }}
-    >
-      <MenuContent>
-        <MenuList>
-          <MenuItem itemId="Account">Account</MenuItem>
-          <MenuItem itemId="Status">Status</MenuItem>
-          <MenuItem itemId="Provider">Provider</MenuItem>
-        </MenuList>
-      </MenuContent>
-    </Menu>
-  );
-
-  const attributeDropdown = (
-    <div ref={attributeContainerRef}>
-      <Popper
-        trigger={attributeToggle}
-        triggerRef={attributeToggleRef}
-        popper={attributeMenu}
-        popperRef={attributeMenuRef}
-        appendTo={attributeContainerRef.current || undefined}
-        isVisible={isAttributeMenuOpen}
-      />
-    </div>
-  );
-
-  return (
-    <Toolbar
-      id="attribute-search-filter-toolbar"
-      clearAllFilters={() => {
-        setSearchValue('');
-        setStatusSelection('');
-        setProviderSelections([]);
-      }}
-    >
-      <ToolbarContent>
-        <ToolbarToggleGroup toggleIcon={<FilterIcon />} breakpoint="xl">
-          <ToolbarGroup variant="filter-group">
-            <ToolbarItem>{attributeDropdown}</ToolbarItem>
-            <ToolbarFilter
-              chips={searchValue !== '' ? [searchValue] : ([] as string[])}
-              deleteChip={() => setSearchValue('')}
-              deleteChipGroup={() => setSearchValue('')}
-              categoryName="Account"
-              showToolbarItem={activeAttributeMenu === 'Account'}
-            >
-              {searchInput}
-            </ToolbarFilter>
-            <ToolbarFilter
-              chips={statusSelection !== '' ? [statusSelection] : ([] as string[])}
-              deleteChip={() => setStatusSelection('')}
-              deleteChipGroup={() => setStatusSelection('')}
-              categoryName="Status"
-              showToolbarItem={activeAttributeMenu === 'Status'}
-            >
-              {statusSelect}
-            </ToolbarFilter>
-            <ToolbarFilter
-              chips={providerSelections}
-              deleteChip={(category, chip) => onProviderMenuSelect(undefined, chip as string)}
-              deleteChipGroup={() => setProviderSelections([])}
-              categoryName="Provider"
-              showToolbarItem={activeAttributeMenu === 'Provider'}
-            >
-              {providerSelect}
-            </ToolbarFilter>
-          </ToolbarGroup>
-        </ToolbarToggleGroup>
-      </ToolbarContent>
-    </Toolbar>
-  );
-};
-
-const ClusterTable: React.FunctionComponent<{
-  searchValue: string;
-  statusFilter: string | null;
-  cloudProviderFilter: string | null;
-  providerSelections;
-}> = ({ searchValue, statusFilter, cloudProviderFilter, providerSelections }) => {
-  const [clusterData, setClusterData] = useState<Cluster[] | []>([]);
-  const [filteredData, setFilteredData] = useState<Cluster[] | []>([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        setLoading(true);
-        const fetchedClusters = await getClusters();
-        setClusterData(fetchedClusters.clusters);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    let filtered = clusterData.filter((cluster) =>
-      cluster.accountName.toLowerCase().includes(searchValue.toLowerCase())
-    );
-
-    if (statusFilter) {
-      filtered = filtered.filter((cluster) => cluster.status === statusFilter);
-    }
-
-    if (providerSelections.length > 0) {
-      filtered = filtered.filter((cluster) => providerSelections.includes(cluster.provider));
-    }
-
-    if (cloudProviderFilter) {
-      filtered = filtered.filter((cluster) => cluster.provider === cloudProviderFilter);
-    }
-
-    setFilteredData(filtered);
-  }, [searchValue, clusterData, statusFilter, providerSelections, cloudProviderFilter]);
-
-  const columnNames = {
-    id: 'ID',
-    name: 'Name',
-    status: 'Status',
-    account: 'Account',
-    cloudProvider: 'Cloud Provider',
-    region: 'Region',
-    nodes: 'Nodes',
-    console: 'Web console',
-  };
-
-  //### Sorting ###
-  // Index of the currently active column
-  const [activeSortIndex, setActiveSortIndex] = React.useState<number | undefined>(0);
-  // sort direction of the currently active column
-  const [activeSortDirection, setActiveSortDirection] = React.useState<'asc' | 'desc' | undefined>('asc');
-  // sort dropdown expansion
-  const getSortableRowValues = (cluster: Cluster): (string | number | null)[] => {
-    const { id, name, status, accountName, provider, region, instanceCount, consoleLink } = cluster;
-    return [id, name, status, accountName, provider, region, instanceCount, consoleLink];
-  };
-
-  // Note that we perform the sort as part of the component's render logic and not in onSort.
-  // We shouldn't store the list of data in state because we don't want to have to sync that with props.
-  let sortedData = filteredData;
-  if (typeof activeSortIndex === 'number' && activeSortIndex !== null) {
-    sortedData = filteredData.sort((a, b) => {
-      const aValue = getSortableRowValues(a)[activeSortIndex];
-      const bValue = getSortableRowValues(b)[activeSortIndex];
-      if (typeof aValue === 'number') {
-        // Numeric sort
-        if (activeSortDirection === 'asc') {
-          return (aValue as number) - (bValue as number);
-        }
-        return (bValue as number) - (aValue as number);
-      } else {
-        // String sort
-        if (activeSortDirection === 'asc') {
-          return (aValue as string).localeCompare(bValue as string);
-        }
-        return (bValue as string).localeCompare(aValue as string);
-      }
-    });
-  }
-
-  const getSortParams = (columnIndex: number): ThProps['sort'] => ({
-    sortBy: {
-      index: activeSortIndex,
-      direction: activeSortDirection,
-      defaultDirection: 'asc', // starting sort direction when first sorting a column. Defaults to 'asc'
-    },
-    onSort: (_event, index, direction) => {
-      setActiveSortIndex(index);
-      setActiveSortDirection(direction);
-    },
-    columnIndex,
-  });
-  //### --- ###
-
-  return (
-    <React.Fragment>
-      {loading ? (
-        <div
-          style={{
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            height: '100vh',
-          }}
-        >
-          <Spinner size="xl" />
-        </div>
-      ) : (
-        <Table aria-label="Sortable table">
-          <Thead>
-            <Tr>
-              <Th sort={getSortParams(0)}>{columnNames.id}</Th>
-              <Th sort={getSortParams(1)}>{columnNames.name}</Th>
-              <Th>{columnNames.status}</Th>
-              <Th sort={getSortParams(3)}>{columnNames.account}</Th>
-              <Th sort={getSortParams(4)}>{columnNames.cloudProvider}</Th>
-              <Th sort={getSortParams(5)}>{columnNames.region}</Th>
-              <Th sort={getSortParams(6)}>{columnNames.nodes}</Th>
-              <Th>{columnNames.console}</Th>
-            </Tr>
-          </Thead>
-          <Tbody>
-            {sortedData.map(cluster => (
-              <Tr key={cluster.id}>
-                <Td dataLabel={columnNames.id}>
-                  <Link to={`/clusters/${cluster.id}`}>{cluster.id}</Link>
-                </Td>
-                <Td dataLabel={columnNames.name}>{cluster.name}</Td>
-                <Td dataLabel={columnNames.status}>{renderStatusLabel(cluster.status)}</Td>
-                <Td dataLabel={columnNames.account}>{cluster.accountName}</Td>
-                <Td dataLabel={columnNames.cloudProvider}>{cluster.provider}</Td>
-                <Td dataLabel={columnNames.region}>{cluster.region}</Td>
-                <Td dataLabel={columnNames.nodes}>{cluster.instanceCount}</Td>
-                <Td dataLabel={columnNames.console}>
-                  <a href={cluster.consoleLink} target="_blank" rel="noopener noreferrer">
-                    Console
-                  </a>
-                </Td>
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      )}
-    </React.Fragment>
-  );
-};
+import { PageSection, PageSectionVariants, Panel, TextContent, Text } from '@patternfly/react-core';
+import React, { useState } from 'react';
+import { useLocation } from 'react-router-dom';
+import ClustersTable from './components/ClustersTable';
+import ClustersTableToolbar from './components/ClustersTableToolbar';
 
 const Clusters: React.FunctionComponent = () => {
   const [searchValue, setSearchValue] = useState<string>('');
@@ -567,15 +21,16 @@ const Clusters: React.FunctionComponent = () => {
       </PageSection>
       <PageSection variant={PageSectionVariants.light} isFilled>
         <Panel>
-          <TableToolbar
+          <ClustersTableToolbar
             searchValue={searchValue}
             setSearchValue={setSearchValue}
-            setStatusSelection={setStatusSelection}
             statusSelection={statusSelection}
-            setProviderSelections={setProviderSelections}
+            setStatusSelection={setStatusSelection}
             providerSelections={providerSelections}
+            setProviderSelections={setProviderSelections}
+            onSearchChange={setSearchValue}
           />
-          <ClusterTable
+          <ClustersTable
             searchValue={searchValue}
             statusFilter={statusSelection}
             cloudProviderFilter={cloudProviderFilter}
